# Precomputed compression mask parts and replacements