@franz1981 2024-12-23
The uops analysis for Intel Skylake is [here](https://uica.uops.info/?code=loop%3A%0D%0A%20%20mov%20%20%20%20%25r8d%2C%25r13d%0D%0A%20%20lea%20%20%20%200x0(%2C%25rbp%2C8)%2C%25r11d%0D%0A%20%20vmovq%20%20%25xmm0%2C%25r10%0D%0A%20%20mov%20%20%20%200x10(%25r10%2C%25r11%2C1)%2C%25rcx%0D%0A%20%20mov%20%20%20%20%25ecx%2C%25r8d%0D%0A%20%20movzbl%20%25r8b%2C%25r11d%0D%0A%20%20mov%20%20%20%200x10(%25r14%2C%25r11%2C4)%2C%25r10d%0D%0A%20%20mov%20%20%20%20%25r8d%2C%25r11d%0D%0A%20%20shr%20%20%20%20%240x18%2C%25r11d%0D%0A%20%20mov%20%20%20%200x10(%25r14%2C%25r11%2C4)%2C%25r9d%0D%0A%20%20mov%20%20%20%20%25r10d%2C%25r11d%0D%0A%20%20shr%20%20%20%20%240x10%2C%25r11d%0D%0A%20%20mov%20%20%20%20%25r9d%2C%25edx%0D%0A%20%20shr%20%20%20%20%240x10%2C%25edx%0D%0A%20%20movslq%20%25r11d%2C%25rdi%0D%0A%20%20movslq%20%25edx%2C%25rax%0D%0A%20%20movzwl%20%25r9w%2C%25edx%0D%0A%20%20shl%20%20%20%20%240x30%2C%25rax%0D%0A%20%20shl%20%20%20%20%240x30%2C%25rdx%0D%0A%20%20movzwl%20%25r10w%2C%25ebx%0D%0A%20%20mov%20%20%20%20%25r8d%2C%25r10d%0D%0A%20%20shr%20%20%20%20%240x10%2C%25r10d%0D%0A%20%20shr%20%20%20%20%240x8%2C%25r8d%0D%0A%20%20movzbl%20%25r10b%2C%25r10d%0D%0A%20%20mov%20%20%20%200x10(%25r14%2C%25r10%2C4)%2C%25r11d%0D%0A%20%20movzbl%20%25r8b%2C%25r8d%0D%0A%20%20mov%20%20%20%200x10(%25r14%2C%25r8%2C4)%2C%25r8d%0D%0A%20%20mov%20%20%20%20%25r11d%2C%25r10d%0D%0A%20%20shr%20%20%20%20%240x10%2C%25r10d%0D%0A%20%20mov%20%20%20%20%25r8d%2C%25r9d%0D%0A%20%20shr%20%20%20%20%240x10%2C%25r9d%0D%0A%20%20movslq%20%25r10d%2C%25r10%0D%0A%20%20movslq%20%25r9d%2C%25r9%0D%0A%20%20shl%20%20%20%20%240x20%2C%25r10%0D%0A%20%20shl%20%20%20%20%240x10%2C%25r9%0D%0A%20%20or%20%20%20%20%20%25r9%2C%25rdi%0D%0A%20%20or%20%20%20%20%20%25r10%2C%25rdi%0D%0A%20%20or%20%20%20%20%20%25rax%2C%25rdi%0D%0A%20%20movzwl%20%25r8w%2C%25r10d%0D%0A%20%20movzwl%20%25r11w%2C%25r11d%0D%0A%20%20shl%20%20%20%20%240x10%2C%25r10%0D%0A%20%20or%20%20%20%20%20%25r10%2C%25rbx%0D%0A%20%20shl%20%20%20%20%240x20%2C%25r11%0D%0A%20%20or%20%20%20%20%20%25r11%2C%25rbx%0D%0A%20%20or%20%20%20%20%20%25rdx%2C%25rbx%0D%0A%20%20pext%20%20%20%25rdi%2C%25rbx%2C%25r11%0D%0A%20%20cmp%20%20%20%200x3c(%25rsp)%2C%25r13d%0D%0A%20%20jae%20%20%20%20buond_check%0D%0A%20%20shr%20%20%20%20%240x20%2C%25rcx%0D%0A%20%20mov%20%20%20%20%25ecx%2C%25r10d%0D%0A%20%20movslq%20%25r13d%2C%25rax%0D%0A%20%20vmovq%20%20%25xmm2%2C%25r8%0D%0A%20%20mov%20%20%20%20%25r11%2C0x10(%25r8%2C%25rax%2C1)%0D%0A%20%20movzbl%20%25r10b%2C%25r11d%0D%0A%20%20mov%20%20%20%200x10(%25r14%2C%25r11%2C4)%2C%25r8d%0D%0A%20%20mov%20%20%20%20%25r10d%2C%25r9d%0D%0A%20%20shr%20%20%20%20%240x18%2C%25r9d%0D%0A%20%20mov%20%20%20%200x10(%25r14%2C%25r9%2C4)%2C%25r11d%0D%0A%20%20mov%20%20%20%20%25r8d%2C%25ecx%0D%0A%20%20shr%20%20%20%20%240x10%2C%25ecx%0D%0A%20%20mov%20%20%20%20%25r11d%2C%25r9d%0D%0A%20%20shr%20%20%20%20%240x10%2C%25r9d%0D%0A%20%20movslq%20%25ecx%2C%25rdx%0D%0A%20%20movslq%20%25r9d%2C%25rcx%0D%0A%20%20movzwl%20%25r11w%2C%25ebx%0D%0A%20%20shl%20%20%20%20%240x30%2C%25rcx%0D%0A%20%20shl%20%20%20%20%240x30%2C%25rbx%0D%0A%20%20movzwl%20%25r8w%2C%25esi%0D%0A%20%20mov%20%20%20%20%25r10d%2C%25r8d%0D%0A%20%20shr%20%20%20%20%240x10%2C%25r8d%0D%0A%20%20shr%20%20%20%20%240x8%2C%25r10d%0D%0A%20%20movzbl%20%25r8b%2C%25r8d%0D%0A%20%20mov%20%20%20%200x10(%25r14%2C%25r8%2C4)%2C%25r11d%0D%0A%20%20movzbl%20%25r10b%2C%25r10d%0D%0A%20%20mov%20%20%20%200x10(%25r14%2C%25r10%2C4)%2C%25r10d%0D%0A%20%20mov%20%20%20%20%25r11d%2C%25r9d%0D%0A%20%20shr%20%20%20%20%240x10%2C%25r9d%0D%0A%20%20mov%20%20%20%20%25r10d%2C%25r8d%0D%0A%20%20shr%20%20%20%20%240x10%2C%25r8d%0D%0A%20%20movslq%20%25r9d%2C%25r9%0D%0A%20%20movslq%20%25r8d%2C%25r8%0D%0A%20%20shl%20%20%20%20%240x20%2C%25r9%0D%0A%20%20shl%20%20%20%20%240x10%2C%25r8%0D%0A%20%20or%20%20%20%20%20%25r8%2C%25rdx%0D%0A%20%20or%20%20%20%20%20%25r9%2C%25rdx%0D%0A%20%20or%20%20%20%20%20%25rcx%2C%25rdx%0D%0A%20%20movzwl%20%25r10w%2C%25r10d%0D%0A%20%20movzwl%20%25r11w%2C%25r11d%0D%0A%20%20shl%20%20%20%20%240x10%2C%25r10%0D%0A%20%20or%20%20%20%20%20%25r10%2C%25rsi%0D%0A%20%20shl%20%20%20%20%240x20%2C%25r11%0D%0A%20%20or%20%20%20%20%20%25r11%2C%25rsi%0D%0A%20%20or%20%20%20%20%20%25rbx%2C%25rsi%0D%0A%20%20pext%20%20%20%25rdx%2C%25rsi%2C%25r9%0D%0A%20%20popcnt%20%25rdi%2C%25r11%0D%0A%20%20mov%20%20%20%20%25r11d%2C%25r10d%0D%0A%20%20sar%20%20%20%20%240x1f%2C%25r10d%0D%0A%20%20shr%20%20%20%20%240x1d%2C%25r10d%0D%0A%20%20add%20%20%20%20%25r11d%2C%25r10d%0D%0A%20%20sar%20%20%20%20%240x3%2C%25r10d%0D%0A%20%20lea%20%20%20%20(%25r10%2C%25r13%2C1)%2C%25r8d%0D%0A%20%20cmp%20%20%20%200x3c(%25rsp)%2C%25r8d%0D%0A%20%20jae%20%20%20%20bound_check%0D%0A%20%20popcnt%20%25rdx%2C%25r11%0D%0A%20%20movslq%20%25r10d%2C%25r8%0D%0A%20%20add%20%20%20%20%25rax%2C%25r8%0D%0A%20%20vmovq%20%20%25xmm2%2C%25rcx%0D%0A%20%20mov%20%20%20%20%25r9%2C0x10(%25rcx%2C%25r8%2C1)%0D%0A%20%20mov%20%20%20%20%25r11d%2C%25r8d%0D%0A%20%20sar%20%20%20%20%240x1f%2C%25r8d%0D%0A%20%20shr%20%20%20%20%240x1d%2C%25r8d%0D%0A%20%20add%20%20%20%20%25r11d%2C%25r8d%0D%0A%20%20sar%20%20%20%20%240x3%2C%25r8d%0D%0A%20%20add%20%20%20%20%25r10d%2C%25r8d%0D%0A%20%20add%20%20%20%20%25r13d%2C%25r8d%0D%0A%20%20inc%20%20%20%20%25ebp%0D%0A%20%20cmp%20%20%20%20(%25rsp)%2C%25ebp%0D%0A%20%20jl%20%20%20%20%20loop%0D%0Abound_check%3A%0D%0A&syntax=asATT&uArchs=SKL&tools=uiCA&tools=IACA3&tools=IACA23&tools=llvm&tools=OSACA&tools=CQA&alignment=0&uiCAHtmlOptions=traceTable&uiCAHtmlOptions=graph&uiCAHtmlOptions=dependencies)