@franz1981 2024-12-25
What I don't like:
1. the 2 potentially overlapping 8 bytes writes: I am not sure what the store buffer will do here. If IIRC these can introduce stalls between stores; it's probably better to do some math and not making it happen. A microbench can easily show it 
2. the 8 lookups happen in two separate sets of 4 load + 1 store, reusing some registers between them. In theory, if registers values are fully rewritten (i.e. no RAW deps) it shouldn't be a problem and register renaming should do the right thing, but still..I want to collect some perf data to make sure of it