@franz1981 2024-12-25
For the former I see https://stackoverflow.com/a/54880249 which seems relevant