@franz1981 2024-12-27
I will play a bit with the idea of removing overlapping writes in the next days - although it cannot fully work (without branches) because of the nature of the algorithm itself and we always risk, in the next iteration to overlap (at best) 4 bytes with the previous write
i.e.we read straight 4 + 4 non special chars and we write 8 + 4 bytes where the last 4 are all zeros.

We could improve the overlapping in the same loop iteration but not across them :/

Right now the limits of the current algorithm are:
- the position where we write the second replacement batch depends by some processing on the first inputs + lookups read
- each batch write depends by the previous iteration result 

So, knowing the reciprocal throughput and latency of the dependency chains (latter very long, former shorter) could be used to guesstimate the IPC - assuming L1d hit for both read/writes